cleanup to trigger rerun
